Move shared React contexts out of components
