+++ packages/react/src/Button/IconButton.tsx
@@ -2,8 +2,9 @@ import React, {forwardRef, type JSX} from 'react'
 import type {IconButtonProps} from './types'
 import type {ForwardRefComponent as PolymorphicForwardRefComponent} from '../utils/polymorphic'
 import {ButtonBase} from './ButtonBase'
-import {TooltipContext, Tooltip} from '../TooltipV2/Tooltip'
-import {TooltipContext as TooltipContextV1} from '../Tooltip/Tooltip'
+import {Tooltip} from '../TooltipV2/Tooltip'
+import {TooltipContext} from '../TooltipV2/TooltipContext'
+import {TooltipContext as TooltipContextV1} from '../Tooltip/TooltipContext'
 import classes from './ButtonBase.module.css'
 import {clsx} from 'clsx'
