+++ packages/react/src/Dialog/DialogContext.ts
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const DialogContext = React.createContext<object | undefined>(undefined)
