+++ packages/react/src/Portal/PortalContext.ts
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const PortalContext = React.createContext<{
+  portalContainerName?: string
+}>({})
