+++ packages/react/src/Portal/index.ts
@@ -1,5 +1,7 @@
 import type {PortalProps} from './Portal'
-import {Portal, registerPortalRoot, PortalContext} from './Portal'
+import {Portal} from './Portal'
+import {PortalContext} from './PortalContext'
+import {registerPortalRoot} from './portalRoot'
 
 export default Portal
 export {registerPortalRoot}
