+++ packages/react/src/RadioGroup/RadioGroupContext.ts
@@ -1,0 +1,7 @@
+import {type ChangeEventHandler, createContext} from 'react'
+
+export const RadioGroupContext = createContext<{
+  disabled?: boolean
+  onChange?: ChangeEventHandler<HTMLInputElement>
+  name: string
+} | null>(null)
