+++ packages/react/src/RadioGroup/index.ts
@@ -1,3 +1,4 @@
-export {default, RadioGroupContext} from './RadioGroup'
+export {default} from './RadioGroup'
+export {RadioGroupContext} from './RadioGroupContext'
 
 export type {RadioGroupProps} from './RadioGroup'
