+++ packages/react/src/Tooltip/TooltipContext.ts
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const TooltipContext = React.createContext<{tooltipId?: string}>({})
