+++ packages/react/src/TooltipV2/TooltipContext.ts
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const TooltipContext = React.createContext<{tooltipId?: string}>({})
